+++ WebApp/Program.cs
@@ -26,11 +26,11 @@ builder.Services.AddIdentity<ApplicationUser, IdentityRole<Guid>>(config =>
     .AddDefaultTokenProviders();
 #endregion
 
-var port = Environment.GetEnvironmentVariable("PORT") ?? "8080";
-builder.WebHost.UseKestrel(options =>
-{
-    options.ListenAnyIP(int.Parse(port));
-});
+//var port = Environment.GetEnvironmentVariable("PORT") ?? "8080";
+//builder.WebHost.UseKestrel(options =>
+//{
+//    options.ListenAnyIP(int.Parse(port));
+//});
 
 // Add Quartz scheduler
 builder.Services.AddSingleton<IScheduler>(provider =>
@@ -69,36 +69,3 @@ app.MapRazorPages();
 #endregion
 
 app.Run();
-
-public class QuartzHostedService : IHostedService
-{
-    private readonly IScheduler _scheduler;
-    private readonly IServiceProvider _serviceProvider;
-
-    public QuartzHostedService(IScheduler scheduler, IServiceProvider serviceProvider)
-    {
-        _scheduler = scheduler;
-        _serviceProvider = serviceProvider;
-    }
-
-    public async Task StartAsync(CancellationToken cancellationToken)
-    {
-        // Start the scheduler
-        await _scheduler.Start(cancellationToken);
-
-        _scheduler.JobFactory = new JobFactory(_serviceProvider);
-
-        // Retrieve the WebAppStartupJobsTrigger from the service provider
-        using (var scope = _serviceProvider.CreateScope())
-        {
-            var jobTrigger = scope.ServiceProvider.GetRequiredService<WebAppStartupJobsTrigger>();
-            await jobTrigger.StartAsync();
-        }
-    }
-
-    public Task StopAsync(CancellationToken cancellationToken)
-    {
-        // Shutdown the scheduler
-        return _scheduler.Shutdown(cancellationToken);
-    }
-}
